analytics/requirements.txt
duckdb==0.10.2
kafka-python==2.0.2
